Implement BindUnmarshaler
This adds the `UnmarshalParam()` function from the `BindUnmarshaler`
interface from echo:
https://github.com/labstack/echo/blob/60072188355b8972c738a041666bd4b05957c135/bind.go#L24

This is useful for unmarshaling form and query values, and since it's
(mostly) a proxy to `UnmarshalText()` I don't see any downsides in
adding it.
The only exception is the `Bool`; I had to add `on` and `off`. I could
add these just for `UnmarshalParam`, but it seemed useful to me to add
it to `UnmarshalText` as well.